@@ -6,27 +6,34 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "eslint",
+    "lint:fix": "eslint --fix"
   },
   "engines": {
     "node": "22.x"
   },
   "dependencies": {
     "@vercel/analytics": "^1.2.2",
-    "@vercel/speed-insights": "^1.0.10",
-    "next": "14.2.32",
-    "react": "^18",
-    "react-dom": "^18"
+    "@vercel/speed-insights": "^1.2.0",
+    "next": "15.5.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
+  "type": "module",
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
+    "@eslint/eslintrc": "^3.3.1",
+    "@next/eslint-plugin-next": "^15.5.2",
+    "@tailwindcss/postcss": "^4.1.12",
+    "@types/node": "^24.3.0",
+    "@types/react": "^18.3.24",
     "@types/react-dom": "^18",
+    "@typescript-eslint/parser": "^8.41.0",
     "autoprefixer": "^10.4.21",
-    "eslint": "^8",
-    "eslint-config-next": "14.0.4",
+    "eslint": "^9.34.0",
+    "eslint-config-next": "15.5.2",
     "postcss": "^8",
-    "tailwindcss": "^3.3.0",
-    "typescript": "^5"
+    "tailwindcss": "^4.1.12",
+    "typescript": "^5.9.2",
+    "typescript-eslint": "^8.41.0"
   }
 }
